requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
uvloop>=0.17.0
websockets>=12.0

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Exchange & Data
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta
pyarrow>=12.0.0
numpy>=1.24.0
requests>=2.31.0
py-clob-client>=0.19.0
web3>=6.0.0

# ML
torch>=2.0.0
scikit-learn>=1.3.0

# Utils
tqdm>=4.65.0
python-dateutil>=2.8.0
